# geox-api/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Mercator API",
    "contact": {
      "email": "duber+ManagementApi@mercator.tech"
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/geofencer/namespaces": {
      "get": {
        "tags": [
          "namespaces"
        ],
        "summary": " Get Namespaces",
        "description": "Return namespaces available to the user.",
        "operationId": "_get_namespaces_geofencer_namespaces_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "ID of the namespace",
              "type": "string",
              "format": "uuid4"
            },
            "name": "id",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Name of the namespace",
              "type": "string"
            },
            "name": "name",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Includeshapes",
              "type": "boolean",
              "default": true
            },
            "name": "includeShapes",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response  Get Namespaces Geofencer Namespaces Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NamespaceResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "namespaces"
        ],
        "summary": " Post Namespaces",
        "description": "Create a new namespace.",
        "operationId": "_post_namespaces_geofencer_namespaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NamespaceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Namespace"
                }
              }
            }
          },
          "404": {
            "description": "Namespace does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/geofencer/namespaces/{namespace_id}": {
      "get": {
        "tags": [
          "namespaces"
        ],
        "summary": " Get Namespaces  Namespace Id",
        "description": "Return a namespace.",
        "operationId": "_get_namespaces__namespace_id_geofencer_namespaces__namespace_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Namespace Id",
              "type": "string",
              "format": "uuid4"
            },
            "name": "namespace_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceResponse"
                }
              }
            }
          },
          "409": {
            "description": "Namespace exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "namespaces"
        ],
        "summary": " Delete Namespaces",
        "description": "Return namespaces available to the user.\n\n- 204: Success with no content",
        "operationId": "_delete_namespaces_geofencer_namespaces__namespace_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Namespace Id",
              "type": "string",
              "format": "uuid4"
            },
            "name": "namespace_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "409": {
            "description": "Namespace exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "namespaces"
        ],
        "summary": " Patch Namespaces",
        "description": "Return namespaces available to the user.",
        "operationId": "_patch_namespaces_geofencer_namespaces__namespace_id__patch",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Namespace to edit",
              "type": "string",
              "format": "uuid4"
            },
            "name": "namespace_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NamespaceUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceResponse"
                }
              }
            }
          },
          "409": {
            "description": "Namespace exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "Namespace does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GeoShapeMetadata": {
        "title": "GeoShapeMetadata",
        "required": [
          "uuid"
        ],
        "type": "object",
        "properties": {
          "uuid": {
            "title": "Uuid",
            "type": "string",
            "format": "uuid4"
          },
          "name": {
            "title": "Name",
            "type": "string",
            "description": "Name of the shape"
          },
          "namespace_id": {
            "title": "Namespace Id",
            "type": "string",
            "format": "uuid4"
          },
          "properties": {
            "title": "Properties",
            "type": "object",
            "description": "Properties of the shape"
          },
          "created_at": {
            "title": "Created At",
            "type": "string",
            "description": "Date and time of creation",
            "format": "date-time"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string",
            "description": "Date and time of most recent updater",
            "format": "date-time"
          }
        },
        "description": "Metadata about a shape.",
        "example": {
          "uuid": "9be2a343-60f5-4324-a5f4-4efe8efeba28",
          "name": "Lake Merritt",
          "properties": {},
          "namespace_id": "cf9f83cd-d454-4abb-b77c-c6d639804618",
          "geojson": {
            "type": "Feature",
            "properties": {},
            "geometry": {
              "type": "Polygon",
              "coordinates": [
                [
                  [
                    -122.26907730102539,
                    37.796084948503236
                  ],
                  [
                    -122.24555969238281,
                    37.796084948503236
                  ],
                  [
                    -122.24555969238281,
                    37.81208947621168
                  ],
                  [
                    -122.26907730102539,
                    37.81208947621168
                  ],
                  [
                    -122.26907730102539,
                    37.796084948503236
                  ]
                ]
              ]
            }
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "Namespace": {
        "title": "Namespace",
        "required": [
          "id",
          "name",
          "slug",
          "created_at",
          "updated_at",
          "is_default"
        ],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string",
            "format": "uuid4"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "pattern": "^[a-z0-9_-]+$",
            "type": "string"
          },
          "properties": {
            "title": "Properties",
            "type": "object"
          },
          "created_at": {
            "title": "Created At",
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string",
            "format": "date-time"
          },
          "is_default": {
            "title": "Is Default",
            "type": "boolean"
          },
          "shapes": {
            "title": "Shapes",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GeoShapeMetadata"
            },
            "description": "\n        List of shape metadata of shapes in the namespace.\n        None means that shape metadata wasn't requested.\n        An empty list means that there are no shapes."
          }
        },
        "description": "Namespace data."
      },
      "NamespaceCreate": {
        "title": "NamespaceCreate",
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "properties": {
            "title": "Properties",
            "type": "object"
          }
        },
        "description": "Data to create a new namespace."
      },
      "NamespaceResponse": {
        "title": "NamespaceResponse",
        "required": [
          "id",
          "name",
          "slug",
          "created_at",
          "updated_at",
          "is_default"
        ],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string",
            "format": "uuid4"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "pattern": "^[a-z0-9_-]+$",
            "type": "string"
          },
          "properties": {
            "title": "Properties",
            "type": "object"
          },
          "created_at": {
            "title": "Created At",
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string",
            "format": "date-time"
          },
          "is_default": {
            "title": "Is Default",
            "type": "boolean"
          },
          "shapes": {
            "title": "Shapes",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GeoShapeMetadata"
            },
            "description": "\n        List of shape metadata of shapes in the namespace.\n        None means that shape metadata wasn't requested.\n        An empty list means that there are no shapes."
          }
        },
        "description": "Response in get namespaces."
      },
      "NamespaceUpdate": {
        "title": "NamespaceUpdate",
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "properties": {
            "title": "Properties",
            "type": "object"
          }
        },
        "description": "Data to update an existing namespace."
      },
...
      "RequestErrorModel": {
        "title": "RequestErrorModel",
        "required": [
          "detail"
        ],
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}